+++ tests/testthat/test-1-get-flights.R
@@ -23,3 +23,17 @@ test_that("standard get_flights (NYC, February 2013)", {
   expect_equal(purrr::map(flights_2, class) %>% unlist(),
                purrr::map(flights_orig, class) %>% unlist())
 })
+
+test_that("arrival get_flights (NYC, February 2013)", {
+  skip_on_cran()
+  skip_if_offline()
+  skip_on_os("windows")
+
+  flights_arr <- get_flights(c("JFK", "LGA", "EWR"), 2013, 2, type = "arrival")
+  expect_gt(nrow(flights_arr), 0)
+  expect_true(all(flights_arr$dest %in% c("JFK", "LGA", "EWR")))
+
+  flights_dep <- get_flights(c("JFK", "LGA", "EWR"), 2013, 2)
+  expect_equal(ncol(flights_arr), ncol(flights_dep))
+  expect_equal(colnames(flights_arr), colnames(flights_dep))
+})